requirements.txt
numpy==1.24.2
pandas==2.0.3
Pillow==10.0.0
scikit_learn==1.3.0
torch==2.0.0
torchvision==0.15.1
transformers==4.30.2
